+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 0.0.1
 
 * Initial release.
+
+## 0.0.2
+* updated intl and implemented new material localization delegate functions
+
+## 0.0.3
+* code optimization
+
+
+## 0.0.4
+
+* added implementation for CupertinoLocalizations.backButtonLabel and  CupertinoLocalizations.cancelButtonLabel